callables: Retain const-correctness for arg->v
Args are generally passed to functions called from a Key or Button. One
of the members, "v", allows passing a `const void *` with the receiving
function determining the intent and casting as appropriate.

There are a few places in the dwm code where we cast away the constness
of this pointer unnecessarily, potentially allowing its mutation. One
case of this is on Linux, where execvp's argv is supposed to be a
`char *const *`, but we just cast away the initial const pointer.

The other case is in setlayout(), but that case is trivial: there's no
issue just casting the const void pointer to a `const Layout *`.